+++ packages/@tailwindcss-vite/src/index.ts
@@ -62,8 +62,12 @@ export default function tailwindcss(opts: PluginOptions = {}): Plugin[] {
 
       let jsResolver = config!.createResolver(config!.resolve)
 
-      customCssResolver = (id: string, base: string) => cssResolver(id, base, true, isSSR)
-      customJsResolver = (id: string, base: string) => jsResolver(id, base, true, isSSR)
+      customCssResolver = async (id: string, base: string) => {
+        let resolved = await cssResolver(id, base, false, isSSR)
+        if (resolved && !resolved.endsWith('.css')) return undefined
+        return resolved
+      }
+      customJsResolver = (id: string, base: string) => jsResolver(id, base, false, isSSR)
     } else {
       type ResolveIdFn = (
         environment: Environment,
@@ -105,8 +109,12 @@ export default function tailwindcss(opts: PluginOptions = {}): Plugin[] {
 
       let jsResolver = createBackCompatIdResolver(env.config, env.config.resolve)
 
-      customCssResolver = (id: string, base: string) => cssResolver(env, id, base, true)
-      customJsResolver = (id: string, base: string) => jsResolver(env, id, base, true)
+      customCssResolver = async (id: string, base: string) => {
+        let resolved = await cssResolver(env, id, base, false)
+        if (resolved && !resolved.endsWith('.css')) return undefined
+        return resolved
+      }
+      customJsResolver = (id: string, base: string) => jsResolver(env, id, base, false)
     }
 
     return new Root(